% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack-package.R
\docType{package}
\name{knapsack-package}
\alias{knapsack-package}
\alias{knapsack}
\title{Lab 6 of Advanced Programming in R}
\description{
This package include following functions:\cr
1: brute_force_knapsack(x, W, fast = FALSE)\cr 
2: knapsack_dynamic(x, W)\cr
3: greedy_knapsack(x, W)\cr
}
