% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{using greedy algorithm to solve knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame contains value and weight of every item}

\item{W}{total weight (max) can hold}
}
\description{
using greedy algorithm to solve knapsack problem
}
