% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{using DP algorithm to solve knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame contains value and weight of every item}

\item{W}{total weight (max) can hold}
}
\description{
using DP algorithm to solve knapsack problem
}
